Update profundizacion_2.py
Cambio de tipo de variable a 'float' a los datos ingresados.

+++ ejercicios_profunidzacion/profundizacion_2.py
@@ -44,8 +44,8 @@ while menu != 'FIN':
         break # Sale del while y el programa termina
     elif menu == '+' or menu == '-' or menu == '*' or menu == '/' or menu == '**': # En caso de que haya ingresado una operación
         print('A continuación ingrese 2 números para realizar una operación.') # Pedido de nros.
-        num_1 = int(input('Ingrese un número para operar: '))
-        num_2 = int(input('Ingrese otro número para operar: '))
+        num_1 = float(input('Ingrese un número para operar: '))
+        num_2 = float(input('Ingrese otro número para operar: '))
         if menu == '+': # Suma
             resultado = num_1 + num_2
         elif menu == '-': # Resta
@@ -59,4 +59,3 @@ while menu != 'FIN':
         print(f'El resultado de operar {num_1} {menu} {num_2} es: {resultado}') # Se imprime el resultado
     else: # En caso de que ingrese un caracter invalido a la variable 'menu'
         print(f'Error, {menu} no es un carácter válido. Por favor ingrese nuevamente.')
-
